PulseCompression: move to reschedule package
