+++ qiskit/pulse/reschedule.py
@@ -16,12 +16,13 @@
 a new Schedule.
 """
 import warnings
+from collections import defaultdict
 
 from typing import List, Optional, Iterable
 
 import numpy as np
 
-from qiskit.pulse import (Acquire, AcquireInstruction, Delay,
+from qiskit.pulse import (Acquire, AcquireInstruction, Delay, Play,
                           InstructionScheduleMap, ScheduleComponent, Schedule)
 from .channels import Channel, AcquireChannel, MeasureChannel, MemorySlot
 from .exceptions import PulseError
@@ -199,3 +200,48 @@ def pad(schedule: Schedule,
         schedule |= Delay(until, channel)
 
     return schedule
+
+
+def compress_pulses(schedules: List[Schedule],
+                    by_schedule: bool = False,
+                    by_channel: bool = False) -> List[Schedule]:
+    """Optimization pass to replace identical pulses.
+
+    Args:
+        schedules (list): Schedules to compress.
+        by_schedule (bool): Compress pulses by schedule.
+        by_channel (bool): Compress pulses by channel.
+
+    Returns:
+        Compressed schedules.
+    """
+
+    def _identifier(sched: str, chan: str) -> str:
+        if not by_channel and not by_schedule:
+            return '__all__'
+
+        pairs = [(chan, by_channel), (sched, by_schedule)]
+        return '_'.join([p[0] for p in pairs if p[1]])
+
+    existing_pulses = defaultdict(list)
+    new_schedules = []
+
+    for schedule in schedules:
+        new_schedule = Schedule(name=schedule.name)
+
+        for time, inst in schedule.instructions:
+            if isinstance(inst, Play):
+                identifier = _identifier(schedule.name, inst.channel.name)
+                if inst.pulse in existing_pulses[identifier]:
+                    idx = existing_pulses[identifier].index(inst.pulse)
+                    identical_pulse = existing_pulses[identifier][idx]
+                    new_schedule |= Play(identical_pulse, inst.channel, inst.name) << time
+                else:
+                    existing_pulses[identifier].append(inst.pulse)
+                    new_schedule |= inst << time
+            else:
+                new_schedule |= inst << time
+
+        new_schedules.append(new_schedule)
+
+    return new_schedules
