+++ releasenotes/notes/compress-indentical-pulses-cd20e82e1525b9f8.yaml
@@ -1,13 +1,13 @@
 ---
 features:
   - |
-    Schedule transformer :func:`qiskit.pulse.transforms.compress_pulses`
+    Schedule transformer :func:`qiskit.pulse.reschedule.compress_pulses`
     performs an optimization pass to reduce the usage of waveform
     memory in hardware by replacing multiple identical instances of
     a pulse in a pulse schedule with a single pulse.
     For example::
 
-        from qiskit.pulse import transforms
+        from qiskit.pulse import reschedule
 
         schedules = []
         for _ in range(2):
@@ -17,4 +17,4 @@ features:
             schedule += Play(SamplePulse([0.0, 0.1]), drive_channel)
             schedules.append(schedule)
 
-        compressed_schedules = transforms.compress_pulses(schedules)
+        compressed_schedules = reschedule.compress_pulses(schedules)
